create deployment in auto mode when a new PR is opened from the github incoming webhook
